refactoring header manager
